move storybook global decorators to preview
