+++ apps/storybook/.storybook/preview.tsx
@@ -1,0 +1,82 @@
+import { ApolloProvider } from "@apollo/client";
+import { MockedProvider } from "@apollo/client/testing";
+import { UserContext, UserContextType } from "@eden/package-context";
+import { apolloClient } from "@eden/package-graphql";
+import { Members } from "@eden/package-graphql/generated";
+import { getMember } from "@eden/package-mock";
+import { DecoratorFn } from "@storybook/react";
+import { SessionProvider } from "next-auth/react";
+import { MemoryRouterProvider } from "next-router-mock/MemoryRouterProvider";
+import * as React from "react";
+import "./global.css";
+
+export const parameters = {
+  apolloClient: {
+    MockedProvider,
+    // any props you want to pass to MockedProvider on every story
+  },
+  actions: { argTypesRegex: "^on[A-Z].*" },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+  options: {
+    storySort: {
+      method: "alphabetical",
+      order: [
+        "Design System",
+        "Cards",
+        ["Project", "User"],
+        "Components",
+        "Containers",
+        "Elements",
+        "Info",
+        "Layout",
+        "Lists",
+        "Modals",
+        "Sections",
+        "Selectors",
+        "Steppers",
+        "*",
+        "Archive",
+      ],
+    },
+  },
+  // controls: { expanded: true },
+  // viewMode: 'docs',
+};
+
+const injectContext: UserContextType = {
+  currentUser: getMember(),
+  memberFound: true,
+  setCurrentUser: (user: Members) => {
+    console.log("setCurrentUser", user);
+    // injectContext.currentUser = user;
+  },
+  refechProfile: () => {},
+  memberServers: [],
+  selectedServer: null,
+  setSelectedServer: () => {},
+};
+
+export const decorators: DecoratorFn[] = [
+  (Story) => (
+    <MemoryRouterProvider>
+      <Story />
+    </MemoryRouterProvider>
+  ),
+  (Story) => (
+    <ApolloProvider client={apolloClient}>
+      <Story />
+    </ApolloProvider>
+  ),
+  (Story) => (
+    <SessionProvider>
+      <UserContext.Provider value={injectContext}>
+        <Story />
+      </UserContext.Provider>
+    </SessionProvider>
+  ),
+];
