+++ packages/ui/src/archive/components/UserProfileMenu/UserProfileMenu.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { UserProfileMenu } from "./UserProfileMenu";
 
@@ -7,7 +6,6 @@ export default {
   title: "Archive/Components/UserProfileMenu",
   component: UserProfileMenu,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof UserProfileMenu>;
 
 const Template: ComponentStory<typeof UserProfileMenu> = (args) => (
