+++ packages/ui/src/archive/containers/ProfileContainer/ProfileContainer.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { ProfileContainer } from "./ProfileContainer";
 
@@ -7,7 +6,6 @@ export default {
   title: "Archive/Containers/ProfileContainer",
   component: ProfileContainer,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof ProfileContainer>;
 
 const Template: ComponentStory<typeof ProfileContainer> = (args) => (
