+++ packages/ui/src/archive/containers/SignUpContainer/SignUpContainer.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { SignUpContainer } from "./";
 import { SignUpProvider } from "./context";
@@ -8,7 +7,6 @@ export default {
   title: "Archive/Containers/SignUpContainer",
   component: SignUpContainer,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof SignUpContainer>;
 
 const Template: ComponentStory<typeof SignUpContainer> = (args) => (
