+++ packages/ui/src/cards/RefineYourProfile/RefineYourProfile.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { RefineYourProfile } from "./RefineYourProfile";
 
@@ -7,7 +6,6 @@ export default {
   title: "Cards/RefineYourProfile",
   component: RefineYourProfile,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof RefineYourProfile>;
 
 const Template: ComponentStory<typeof RefineYourProfile> = () => (
