+++ packages/ui/src/cards/user/UserProfileCard/UserProfileCard.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { UserProfileCard } from "./UserProfileCard";
 
@@ -7,7 +6,6 @@ export default {
   title: "Cards/User/UserProfileCard",
   component: UserProfileCard,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof UserProfileCard>;
 
 const Template: ComponentStory<typeof UserProfileCard> = (args) => (
