+++ packages/ui/src/components/SendMessageToChampion/SendMessageToChampion.stories.tsx
@@ -1,6 +1,5 @@
 import { getMember } from "@eden/package-mock";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { SendMessageToChampion } from "./SendMessageToChampion";
 
@@ -8,7 +7,6 @@ export default {
   title: "Components/SendMessageToChampion",
   component: SendMessageToChampion,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof SendMessageToChampion>;
 
 const Template: ComponentStory<typeof SendMessageToChampion> = (args) => {
