+++ packages/ui/src/components/SendMessageToUser/SendMessageToUser.stories.tsx
@@ -1,6 +1,5 @@
 import { getMember } from "@eden/package-mock";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { SendMessageToUser } from "./SendMessageToUser";
 
@@ -8,7 +7,6 @@ export default {
   title: "Components/SendMessageToUser",
   component: SendMessageToUser,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof SendMessageToUser>;
 
 const Template: ComponentStory<typeof SendMessageToUser> = (args) => {
