+++ packages/ui/src/lists/ProjectChampionList/ProjectChampionList.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { ProjectChampionList } from "./ProjectChampionList";
 
@@ -7,7 +6,6 @@ export default {
   title: "Lists/ProjectChampionList",
   component: ProjectChampionList,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof ProjectChampionList>;
 
 const Template: ComponentStory<typeof ProjectChampionList> = () => {
