+++ packages/ui/src/modals/ApplyByRoleModal/ApplyByRoleModal.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { ApplyByRoleModal } from "./ApplyByRoleModal";
 
@@ -7,7 +6,6 @@ export default {
   title: "Modals/ApplyByRoleModal",
   component: ApplyByRoleModal,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof ApplyByRoleModal>;
 
 const Template: ComponentStory<typeof ApplyByRoleModal> = (args) => (
