+++ packages/ui/src/modals/SendMessageToUserModal/SendMessageToUserModal.stories.tsx
@@ -1,6 +1,5 @@
 import { getMember } from "@eden/package-mock";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { SendMessageToUserModal } from "./SendMessageToUserModal";
 
@@ -8,7 +7,6 @@ export default {
   title: "Modals/SendMessageToUserModal",
   component: SendMessageToUserModal,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof SendMessageToUserModal>;
 
 const Template: ComponentStory<typeof SendMessageToUserModal> = (args) => {
