+++ packages/ui/src/selectors/SubmenuSelector/SubmenuSelector.stories.tsx
@@ -1,5 +1,4 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { CurrentUserDecorator } from "storybook/.storybook/decorator";
 
 import { SubmenuSelector } from "./SubmenuSelector";
 
@@ -7,7 +6,6 @@ export default {
   title: "Components/SubmenuSelector",
   component: SubmenuSelector,
   argTypes: {},
-  decorators: [CurrentUserDecorator],
 } as ComponentMeta<typeof SubmenuSelector>;
 
 const Template: ComponentStory<typeof SubmenuSelector> = (args) => (
